requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
filelock==3.13.1
fsspec==2024.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python-headless==4.11.0.86
pillow==11.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pytorch-triton-rocm==3.3.1
PyYAML==6.0.2
scipy==1.16.0
setuptools==70.2.0
simsimd==6.4.9
stringzilla==3.12.5
sympy==1.13.3
torch==2.7.1+rocm6.3
torchaudio==2.7.1+rocm6.3
torchvision==0.22.1+rocm6.3
typing-inspection==0.4.1
typing_extensions==4.12.2